DynamicElectricSheath

Documentation for DynamicElectricSheath

DynamicElectricSheath.Discretization

DynamicElectricSheath.Physics

DynamicElectricSheath.E0

DynamicElectricSheath.advection!

DynamicElectricSheath.compute_charge!

DynamicElectricSheath.compute_current

DynamicElectricSheath.compute_e!

DynamicElectricSheath.fe_0

DynamicElectricSheath.fi_0

DynamicElectricSheath.mask

DynamicElectricSheath.Discretization — Type

struct Discretization

Grid in velocity corresponds to the electronic one which contains the support of boths ions and electron density function

Nx::Int64

Number of points of the space mesh

Nv::Int64

Number of points of the speed mesh

dx::Float64

Space step

dv::Float64

Speed step

CFL_x::Float64

CFL in x

CFL_v::Float64

CFL in v

Nt::Int64

Number of time steps

dt::Float64

Time step

DynamicElectricSheath.Physics — Type

struct Physics

ν::Float64

Collision frequency Default: 20.0

λ::Float64

Debye length Default: 0.5

T::Float64

Time horizon Default: 10.0

μ::Float64

Thermal speed Default: 0.5

xmin::Float64

Lower bound space Default: -1.0

xmax::Float64

Upper bound space Default: 1.0

vmin::Float64

Lower bound speed Default: -10.0

vmax::Float64

Uupper bound speed Default: 10.0

DynamicElectricSheath.E0 — Method

E0(x)

Electric field

DynamicElectricSheath.advection! — Method

advection!(
    fi,
    fe,
    vv_plus,
    vv_minus,
    EE_plus,
    EE_minus,
    nu,
    mu,
    dx,
    dv,
    dt
)

\[\frac{d f_i}{dt} + v \frac{d f_i}{dx} + E \frac{d f_i}{dv} = \nu * f_e\]

\[\frac{d f_e}{dt} + v \frac{d f_e}{dx} + \frac{1}{\mu} E \frac{d f_e}{dv} = 0\]

DynamicElectricSheath.compute_charge! — Method

compute_charge!(rho, f, dv)

compute charge density

DynamicElectricSheath.compute_current — Method

compute_current(fi, fe, vv, dv)

compute current at boundaries : rectangle integration of

\[\int_{v} v * (fi(t,\pm 1,v) - fe(t,\pm 1,v)) dv\]

DynamicElectricSheath.compute_e! — Method

compute_e!(EE, rho, lambda, J_l, J_r, dx, dt)

Update electric field : use formula

\[E(x) = 0.5*(E(1) + E(-1)) + \nu/2 * ( \int_{-1}^{x} rho(y)dy - \int_{x}^{1} \rho(y)dy )\]

with $E(t^{n+1},1) + E(t^{n+1},-1)$ approximated by $E(t^n,1) + E(t^n,-1) + dt/\lambda^2 * (J_l + J_r)$ (Ampère boundary conditions)

DynamicElectricSheath.fe_0 — Method

fe_0(x, v; μ)

Electrons distribution

DynamicElectricSheath.fi_0 — Method

fi_0(x, v; μ)

Ions distribution

DynamicElectricSheath.mask — Method

mask(x)

Initial mask